[PyTorch] Move test_grouped_gemm_quant_cute to test_grouped_linear
The test validates that grouped_gemm_quant_wrapper_sm100 (a cuTE DSL
kernel internal to the grouped MLP fusion) matches MXFP8 quantizer
output. It does not exercise the op-fuser infrastructure at all, so
it belongs in test_grouped_linear.py alongside the other grouped-MLP-
specific tests.

Also removes the now-unused `import transformer_engine_torch as tex`
from test_fusible_ops.py.
