+++ tests/pytorch/test_fusible_ops.py
@@ -40,7 +40,6 @@ from transformer_engine.pytorch import (
     is_bf16_available,
 )
 from transformer_engine.pytorch.tensor.grouped_tensor import GroupedTensor
-import transformer_engine_torch as tex
 
 # Import utility functions
 from utils import (
@@ -4302,144 +4301,3 @@ class TestTrainingLoops:
             quantization=quantization,
             recipe=recipe,
         )
-
-
-def test_grouped_gemm_quant_cute_matches_mxfp8_quantized() -> None:
-    if not mxfp8_available:
-        pytest.skip(reason_for_no_mxfp8)
-    if torch.cuda.get_device_capability() < (10, 0):
-        pytest.skip("Requires SM100+ for grouped GEMM quant kernel.")
-
-    try:
-        from cudnn import grouped_gemm_quant_wrapper_sm100  # pylint: disable=no-name-in-module
-    except ImportError as exc:
-        pytest.skip(f"grouped_gemm_quant_wrapper_sm100 unavailable: {exc}")
-
-    device = torch.device("cuda")
-    dtype = torch.bfloat16 if is_bf16_available() else torch.float16
-    num_groups = 4
-    m = 256
-    n = 512
-    k = 512
-    total_m = num_groups * m
-    split_sizes = torch.full((num_groups,), m, device=device, dtype=torch.int64)
-
-    q = MXFP8Quantizer(fp8_dtype=te.DType.kFloat8E4M3, rowwise=True, columnwise=False)
-    q.optimize_for_gemm = False
-
-    torch.manual_seed(0)
-    a_full = torch.randn(total_m, k, device=device, dtype=dtype)
-    weights = [torch.randn(n, k, device=device, dtype=dtype) for _ in range(num_groups)]
-
-    grouped_a = tex.group_quantize(a_full, q, num_groups, split_sizes)
-    a_groups = grouped_a.split_into_quantized_tensors()
-    b_groups = [q(w) for w in weights]
-
-    # Reference GEMM on dequantized tensors.
-    ref = torch.empty((total_m, n), device=device, dtype=torch.float32)
-    start = 0
-    for group_idx in range(num_groups):
-        end = start + m
-        a_deq = a_groups[group_idx].dequantize(dtype=torch.float32)
-        b_deq = b_groups[group_idx].dequantize(dtype=torch.float32)
-        ref[start:end, :] = a_deq @ b_deq.t()
-        start = end
-    ref = ref.to(dtype=torch.bfloat16).to(torch.float32)
-
-    # Allocate empty input tensors needed for cuTE DSL kernel
-    padded_offsets = torch.tensor(
-        [m * (i + 1) for i in range(num_groups)],
-        dtype=torch.int32,
-        device=device,
-    )
-    inputs = {
-        "a_tensor": torch.empty(1, total_m, k, dtype=torch.float8_e4m3fn, device=device).permute(
-            1, 2, 0
-        ),
-        "b_tensor": torch.empty(num_groups, n, k, dtype=torch.float8_e4m3fn, device=device).permute(
-            1, 2, 0
-        ),
-        "sfa_tensor": torch.empty(
-            1,
-            total_m // 128,
-            k // 128,
-            32,
-            4,
-            4,
-            dtype=torch.float8_e8m0fnu,
-            device=device,
-        ).permute(3, 4, 1, 5, 2, 0),
-        "sfb_tensor": torch.empty(
-            num_groups,
-            n // 128,
-            k // 128,
-            32,
-            4,
-            4,
-            dtype=torch.float8_e8m0fnu,
-            device=device,
-        ).permute(3, 4, 1, 5, 2, 0),
-        "alpha_tensor": torch.empty(num_groups, dtype=torch.float32, device=device),
-        "prob_tensor": torch.empty(total_m, 1, 1, dtype=torch.float32, device=device),
-        "padded_offsets_tensor": padded_offsets,
-    }
-    # Overwrite inputs with quantized data/scales from MXFP8 quantizer.
-    a_data = grouped_a.rowwise_data.view(total_m, k).view(dtype=torch.float8_e4m3fn)
-    a_data = a_data.unsqueeze(0).permute(1, 2, 0).contiguous()
-    inputs["a_tensor"].copy_(a_data)
-
-    a_scales = grouped_a.scale_inv.view(dtype=torch.float8_e8m0fnu)
-    a_scales = a_scales.view(1, total_m // 128, 4, 32, k // 128, 4)
-    a_scales = a_scales.permute(0, 1, 4, 3, 2, 5).contiguous()
-    a_scales = a_scales.permute(3, 4, 1, 5, 2, 0).contiguous()
-    inputs["sfa_tensor"].copy_(a_scales)
-
-    b_data = torch.cat([w._rowwise_data.reshape(-1) for w in b_groups])
-    b_data = b_data.view(dtype=torch.float8_e4m3fn)
-    b_data = b_data.view(num_groups, n, k).permute(1, 2, 0).contiguous()
-    inputs["b_tensor"].copy_(b_data)
-
-    b_scales = torch.cat([w._rowwise_scale_inv for w in b_groups])
-    b_scales = b_scales.view(dtype=torch.float8_e8m0fnu)
-    b_scales = b_scales.view(num_groups, n // 128, 4, 32, k // 128, 4)
-    b_scales = b_scales.permute(0, 1, 4, 3, 2, 5).contiguous()
-    b_scales = b_scales.permute(3, 4, 1, 5, 2, 0).contiguous()
-    inputs["sfb_tensor"].copy_(b_scales)
-
-    inputs["alpha_tensor"].fill_(1.0)
-    inputs["prob_tensor"].fill_(1.0)
-
-    cute_out = grouped_gemm_quant_wrapper_sm100(
-        a_tensor=inputs["a_tensor"],
-        b_tensor=inputs["b_tensor"],
-        sfa_tensor=inputs["sfa_tensor"],
-        sfb_tensor=inputs["sfb_tensor"],
-        padded_offsets=inputs["padded_offsets_tensor"],
-        alpha_tensor=inputs["alpha_tensor"],
-        norm_const_tensor=None,
-        prob_tensor=inputs["prob_tensor"],
-        acc_dtype=torch.float32,
-        d_dtype=torch.bfloat16,
-        cd_major="n",
-        sf_vec_size=32,
-        discrete_col_sfd=True,
-        current_stream=None,
-    )
-
-    if isinstance(cute_out, dict):
-        outputs = cute_out
-    else:
-        d_tensor, d_col_tensor, amax_tensor, sfd_row_tensor, sfd_col_tensor = cute_out
-        outputs = {
-            "d_tensor": d_tensor,
-            "d_col_tensor": d_col_tensor,
-            "amax_tensor": amax_tensor,
-            "sfd_row_tensor": sfd_row_tensor,
-            "sfd_col_tensor": sfd_col_tensor,
-        }
-
-    d_cute = outputs["d_tensor"]
-    if d_cute.dim() == 3:
-        d_cute = d_cute.squeeze(-1)
-    tols = dtype_tols(torch.bfloat16)
-    assert_close(d_cute[:total_m].float(), ref, **tols)
